requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.7.0
PyPDF2==1.24
PyYAML==3.11
Rtree==0.8.2
SQLAlchemy==0.8.7
Wand==0.4.0
Werkzeug==0.10.4
amqp-dispatcher==0.1.1
argparse==1.2.1
boto==2.30.0
coloredlogs==0.8
coverage==3.7.1
coveralls==0.5
docopt==0.6.2
dsnparse==0.1.3
elasticsearch==1.4.0
fuzzywuzzy==0.5.0
gevent==1.0.1
greenlet==0.4.7
haigha==0.7.3
humanfriendly==1.14
itsdangerous==0.24
lxml==3.4.2
newrelic==2.50.0.39
nose==1.3.3
oauthlib==0.7.2
pika==0.9.14
pip-accel==0.13.5
python-Levenshtein==0.12.0
python-dateutil==2.4.0
python-twitter==2.3
qrcode==5.1
raven==5.1.1
requests==2.5.1
requests-oauthlib==0.5.0
six==1.9.0
urllib3==1.10.4
wsgiref==0.1.2
zbar==0.10
